Update information entity
