+++ taotie/sources/github.py
@@ -72,9 +72,9 @@ class GithubTrends(BaseSource):
                         type="github-repo",
                         datetime_str=get_datetime(),
                         id=repo_name,
-                        repo_url=repo_url,
+                        uri=repo_url,
+                        conent=repo_readme,
                         repo_desc=repo_desc,
-                        repo_readme=repo_readme,
                         repo_lang=repo_lang,
                         repo_star=repo_star,
                         repo_fork=repo_fork,
